Code motion refactorin in storage::kv
