+++ src/storage/kv/stats.rs
@@ -1,0 +1,163 @@
+// Copyright 2019 TiKV Project Authors. Licensed under Apache-2.0.
+
+use engine::{CF_DEFAULT, CF_LOCK, CF_WRITE};
+use tikv_util::collections::HashMap;
+use tikv_util::worker::FutureScheduler;
+use kvproto::kvrpcpb::{ScanDetail, ScanInfo};
+
+use crate::raftstore::store::PdTask;
+
+const STAT_TOTAL: &str = "total";
+const STAT_PROCESSED: &str = "processed";
+const STAT_GET: &str = "get";
+const STAT_NEXT: &str = "next";
+const STAT_PREV: &str = "prev";
+const STAT_SEEK: &str = "seek";
+const STAT_SEEK_FOR_PREV: &str = "seek_for_prev";
+const STAT_OVER_SEEK_BOUND: &str = "over_seek_bound";
+
+/// Statistics collects the ops taken when fetching data.
+#[derive(Default, Clone, Debug)]
+pub struct CFStatistics {
+    // How many keys that's effective to user. This counter should be increased
+    // by the caller.
+    pub processed: usize,
+    pub get: usize,
+    pub next: usize,
+    pub prev: usize,
+    pub seek: usize,
+    pub seek_for_prev: usize,
+    pub over_seek_bound: usize,
+    pub flow_stats: FlowStatistics,
+}
+
+#[derive(Default, Debug, Clone)]
+pub struct FlowStatistics {
+    pub read_keys: usize,
+    pub read_bytes: usize,
+}
+
+// Reports flow statistics to outside.
+pub trait FlowStatsReporter: Send + Clone + Sync + 'static {
+    // Reports read flow statistics, the argument `read_stats` is a hash map
+    // saves the flow statistics of different region.
+    // TODO: maybe we need to return a Result later?
+    fn report_read_stats(&self, read_stats: HashMap<u64, FlowStatistics>);
+}
+
+impl FlowStatsReporter for FutureScheduler<PdTask> {
+    fn report_read_stats(&self, read_stats: HashMap<u64, FlowStatistics>) {
+        if let Err(e) = self.schedule(PdTask::ReadStats { read_stats }) {
+            error!("Failed to send read flow statistics"; "err" => ?e);
+        }
+    }
+}
+
+impl FlowStatistics {
+    pub fn add(&mut self, other: &Self) {
+        self.read_bytes = self.read_bytes.saturating_add(other.read_bytes);
+        self.read_keys = self.read_keys.saturating_add(other.read_keys);
+    }
+}
+
+impl CFStatistics {
+    #[inline]
+    pub fn total_op_count(&self) -> usize {
+        self.get + self.next + self.prev + self.seek + self.seek_for_prev
+    }
+
+    pub fn details(&self) -> [(&'static str, usize); 8] {
+        [
+            (STAT_TOTAL, self.total_op_count()),
+            (STAT_PROCESSED, self.processed),
+            (STAT_GET, self.get),
+            (STAT_NEXT, self.next),
+            (STAT_PREV, self.prev),
+            (STAT_SEEK, self.seek),
+            (STAT_SEEK_FOR_PREV, self.seek_for_prev),
+            (STAT_OVER_SEEK_BOUND, self.over_seek_bound),
+        ]
+    }
+
+    pub fn add(&mut self, other: &Self) {
+        self.processed = self.processed.saturating_add(other.processed);
+        self.get = self.get.saturating_add(other.get);
+        self.next = self.next.saturating_add(other.next);
+        self.prev = self.prev.saturating_add(other.prev);
+        self.seek = self.seek.saturating_add(other.seek);
+        self.seek_for_prev = self.seek_for_prev.saturating_add(other.seek_for_prev);
+        self.over_seek_bound = self.over_seek_bound.saturating_add(other.over_seek_bound);
+        self.flow_stats.add(&other.flow_stats);
+    }
+
+    pub fn scan_info(&self) -> ScanInfo {
+        let mut info = ScanInfo::default();
+        info.set_processed(self.processed as i64);
+        info.set_total(self.total_op_count() as i64);
+        info
+    }
+}
+
+#[derive(Default, Clone, Debug)]
+pub struct Statistics {
+    pub lock: CFStatistics,
+    pub write: CFStatistics,
+    pub data: CFStatistics,
+}
+
+impl Statistics {
+    pub fn total_op_count(&self) -> usize {
+        self.lock.total_op_count() + self.write.total_op_count() + self.data.total_op_count()
+    }
+
+    pub fn total_processed(&self) -> usize {
+        self.lock.processed + self.write.processed + self.data.processed
+    }
+
+    pub fn details(&self) -> [(&'static str, [(&'static str, usize); 8]); 3] {
+        [
+            (CF_DEFAULT, self.data.details()),
+            (CF_LOCK, self.lock.details()),
+            (CF_WRITE, self.write.details()),
+        ]
+    }
+
+    pub fn add(&mut self, other: &Self) {
+        self.lock.add(&other.lock);
+        self.write.add(&other.write);
+        self.data.add(&other.data);
+    }
+
+    pub fn scan_detail(&self) -> ScanDetail {
+        let mut detail = ScanDetail::default();
+        detail.set_data(self.data.scan_info());
+        detail.set_lock(self.lock.scan_info());
+        detail.set_write(self.write.scan_info());
+        detail
+    }
+
+    pub fn mut_cf_statistics(&mut self, cf: &str) -> &mut CFStatistics {
+        if cf.is_empty() {
+            return &mut self.data;
+        }
+        match cf {
+            CF_DEFAULT => &mut self.data,
+            CF_LOCK => &mut self.lock,
+            CF_WRITE => &mut self.write,
+            _ => unreachable!(),
+        }
+    }
+}
+
+#[derive(Default, Debug)]
+pub struct StatisticsSummary {
+    pub stat: Statistics,
+    pub count: u64,
+}
+
+impl StatisticsSummary {
+    pub fn add_statistics(&mut self, v: &Statistics) {
+        self.stat.add(v);
+        self.count += 1;
+    }
+}
